refactor: transfer of shared route to shared api folder
